Fixes router + ResourceContainer component
